pyproject.toml
[project]
name = "fin-trade"
version = "0.1.0"
description = ""
authors = [
    {name = "digital-thinking",email = "github@digital-thinking.de"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "yfinance>=0.2.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pyyaml>=6.0",
    "plotly>=5.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "fastapi (>=0.128.6,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "websockets (>=16.0,<17.0)",
    "psutil>=5.9.0",
    "python-multipart>=0.0.6",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
packages = [{include = "fin_trade", from = "backend"}]


[tool.poetry.scripts]
fin-trade = "backend.main:main"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
httpx = "^0.28.1"
pytest-asyncio = "^1.3.0"


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short --asyncio-mode=auto"
markers = [
    "asyncio: mark test as async",
]


[tool.coverage.run]
source = ["backend"]
omit = []

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
fail_under = 35
show_missing = true
